Update Dockerfile for VSCode notebook compatibility
Two small changes to Dockerfile. Updated the Openstudio args at the top to the current version of Openstudio and to use the current Openstudio image. I see there's already a pull request to update the docker-compose file so I left that alone. The other change is that I modified the final command to disable check xsrf. Disabling the xsrf check allows VSCode to connect to the remote notebook and run cells without an error. Since the server already runs the notebook without authentication and is meant to be run locally I don't think this is an issue.